// offscreen_service/src/types.ts
export enum MessageOrigin {
  OFFSCREEN_TX_TO_BACKGROUND = "offscreen_tx_to_background",
  OFFSCREEN_SCANNER_TO_BACKGROUND = "offscreen_scanner_to_background",
  BACKGROUND_TO_OFFSCREEN_TX = "background_to_offscreen_tx",
  BACKGROUND_TO_OFFSCREEN_SCANNER = "background_to_offscreen_scanner",
}

export interface OffscreenMessagePayload<T = any> {
  error: null | string;
  data: T | null;
}

export enum OffscreenMessageType {
  ERROR = "error",
  RESPONSE = "response",
}

export interface OffscreenMessage<T = any> {
  type: OffscreenMessageType;
  origin: MessageOrigin;
  payload: OffscreenMessagePayload<T>;
}

export enum OffscreenMethod {
  SEND_TX = "send_tx",
  DEPLOY = "deploy",
  IS_SENDING_TX = "is_sending_tx",
  SCANNER_ENCRYPT_REGISTRATION = "scanner_encrypt_registration",
  SCANNER_PING = "scanner_ping",
}

export interface ScannerEncryptRegistrationPayload {
  publicKey: string;
  viewKey: string;
  start: number;
}

export interface ScannerEncryptRegistrationResult {
  ciphertext: string;
}

export interface BackgroundMessage<T = any> {
  type: OffscreenMethod;
  origin: MessageOrigin;
  payload: T;
}